Primality Test
The Primality Test determines whether a number (n) is a prime number.
The algorithm divides the numbers that are less than n.
If there is a number that is divisible by n then we return false,
meaning the input number is not a prime number.

If there is a number that is only divisible by 1 then we return true,
meaning the input number is a prime number.

Best case: n = 1 OR n = 2. We have O(1) Constant Time
Average case: This is when n is an average number. eg. n = 8. We get O(n), Linear Time
Worst case: This is when n is a very large prime number,
eg: n = 27, 277. We have O(n) Linear Time