Improve Primality Test
A new solution that uses is more efficient than the the previous solution.
This solution only iterates until the square root of n, which means there
are less loops.

This solution works as a number that is not a prime number has at least one
factor that is smaller or equal to the square root of the number.

The Time Complexity for this solution is: O(sqrt(n)).