Find smallest number in an array
`getMin` finds the smallest number in a given array.
The algorithm keeps track of the current minimum number.
The `currentMin` changes through a comparison statement.

Time Complexity: O(n), Linear Complexity.
This is because all numbers will need to go through the for loop
and the comparison statement. The efficiency of the algorithm depends
on the number of values in the input array.

The function `findMin` uses `Math.min()`. This algorithm does work
however, the time complexity is somewhat hidden in the built in `Math`
function.