Find the even and odd numbers
This algorithm uses the modulo to determine if the input (n) is even or odd.

Time Complexity: O(1), Constant time.
This is because the modulo function will only ever run once, no matter
how large the n is.
All cases, best, worst and average, will have a time complexity of O(1).