Algorithm: Power of two
This function takes in n and determines whether n is a power of two.

It is not enough to just use a modulo to solve this problem, eg: n % 2 !== 0, as in the case when n = 20, it is divisible by two but it is not a power of two. Thus, rendering an incorrect answer.
Hence, the need for the while loop. We do however need the modulo within the while loop to prevent an infinite loop.
For example, if n = 10, we would enter the while loop, the next number would be 5, then 2.5, then 1.25 and so on but never one. So the modulo is required to escape this loop.

Time Complexity:
Logarithmic Time O(log n)

Best case would be a number that is not divisible by two, eg: n =3, this would be a Linear Time, O(1).
Worst case would be a very large number that is divisible by two as it needs to enter the while loop several times, eg: 1,125,899,906,842,624. This would produce a Logarithmic Time Complexity. This is because for a small number that is divisible by two, eg: n = 32, it will only take 5 iterations, wheras a large number like 1,125,899,906,842,624 will take 50 interations. So the iterations are not correlated with the magnitude of n.