Algorithms: Power of Two Optimised
This solution uses the bitwise operator `&`
This solution compares the binary of the numbers to determine of `n` is a power of two.
Time Complexity: O(1), Constant Time.

Explanation:
When a number is a power of two they only every have one bit, eg:
2 => 0010
4 => 0100
8 => 1000
...etc

To compare numbers in binary, we check the places, if there is a 0 present that amounts to a 0. Eg:
4 => 0100
8 => 1000
=> 0000
So here `4 & 8` equals 0.

If we compare `4 & 7`, we get 4.
4 => 0100
7 => 0111
=> 0100

If we use the new formula, `n & (n-1)` we will always get 0 for a number that is a power of two. Eg:
`4 & 3`, equals 0.
4 => 0100
3 => 0011
=> 0000