Recursion: Factorial
The factorial algorithm has been reworked to use recursion.
The main take away here is recursive algorithms always require a base case.

In this case the base case is the `if` statement. This prevents an infinite loop as recursion is essentially a function that calls itself.

Time Complexity: O(n), Linear Time. The time complexity remains unchanged from the previous solution. The only real advantage of using recursion here is that there is less code written.