Linear Search Algorithm
The linear search algorithm is a simple algorithm that looks at each element.
Once the element is found, the index of the element is returned.

Note, this algorithm only works on primitives, such as numbers and strings.
It does not work on objects are references, so two objects are represented in two different places in memory.
So the objects will never be equal.

Time complexity:
- Best case: O(1), Constant time. This would be if the element we're looking for is the first element.
- Average case: O(n), Linear time. When we don't know where the element could be in the array.
- Worst case: O(n). When the element we're looking for is the last element.
We always take the worst case, so the time complexity is O(n), Linear time.