Linear Search for Objects
This function allows for linear search of objects.

It takes in a comparator function, which makes the function more dynamic.
The comparator function allows you to pass in the key, in this case 'name'.
This means we would need to know the key we want to pass in and compare.

The comparator function finds the key and compares the values of the keys,