findIndex()
JavaScript has a built in method called `findIndex()` that returns the index of an element if found.
If the element is not found, the method returns -1, which is why the test has been changed.

More info on the method: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/findIndex