Quicksort
This algorithm is divides and conquers.
Here we divide the array into smaller chunks and sort them recursively.
This array uses a pivot element and partitions the array into two sub-arrays,
one sub-array holds the elements that are smaller than the pivot and
another sub-array holds the elements that are larger than the pivot.
We then take each sub-array and sort those (recursively).

Time complexity: Best case and average case, the arrays are not sorted. O(n * log n)
Worst case, the arrays are sorted. O(n^2)

The reason the worst case is when the array is already sorted is because we do not take advantage
of the splitting into sub-arrays.